Finish apply blocks updates
